Added Structs and Error.go
Error.go just makes it so that I don't need to keep typing out that if
err != nil thing. Makes it such that its Enil(err) now. Easy.

I added three main components: xl, yt, and sorter. xl.go is just an
interface for any excel files that will come in and out of the program.
yt.go is an interface for any YouTube videos that need to be processed.
sorter.go uses both of these interfaces to sort the data and find which
ones are necessary and which ones aren't, then exports that data to a
slice which is made of structs called RelevantData.

TODO: Export sorted data to excel sheet. Write to excel sheet.